isGuessRight() Implementation changed the function declaration for guessChange)() to isGuessRight(0), that takes in 2 parameters. the original charArr and the copy of it with the randomly changed character.
Finished the implementation of the function to return a boolean whether or not the guessed locaation of the character
change was correct or not.

This will be called in playGame() to detect win/lose conditions. playGame()'s logic will be implemented next.

The whole program has been tested and debugged thus far, and all functions work as intended.